+++ lib/grepfruit/programmatic_search.rb
@@ -1,7 +1,7 @@
 module Grepfruit
   class ProgrammaticSearch < Search
     def execute
-      raise ArgumentError, "directory '#{path}' does not exist." unless File.exist?(path)
+      raise ArgumentError, "path '#{path}' does not exist." unless File.exist?(path)
 
       build_result_hash(execute_search)
     end