+++ spec/grepfruit/cli_search_spec.rb
@@ -280,10 +280,10 @@ RSpec.describe Grepfruit::CliSearch do
       it { is_expected.to include("Error: Jobs must be at least 1") }
     end
 
-    context "when searching non-existent directory" do
+    context "when searching non-existent path" do
       subject { `./exe/grepfruit search -r 'TODO' ./nonexistent` }
 
-      it { is_expected.to include("Error: Directory") }
+      it { is_expected.to include("Error: Path") }
       it { is_expected.to include("does not exist") }
     end
   end
@@ -328,7 +328,7 @@ RSpec.describe Grepfruit::CliSearch do
         json = JSON.parse(subject)
         expect(json["search"]).to include(
           "pattern" => "/TODO/",
-          "directory" => File.expand_path("./spec/test_dataset")
+          "path" => File.expand_path("./spec/test_dataset")
         )
         expect(json["search"]["timestamp"]).to match(/\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}/)
       end